FIX deadlock sending data to handlers
If the channel is created inside the goroutine, the main function's
send via:
        handlermap[id].Input <- m

ends up somewhere where the handler's read via:
        select {
        case m := <-h.Input:
                h.foo(m)
        }

does not actually receive it. At this point all go routines block
on channels, and nothing moves any further. But this is not detected
by the runtime as deadlock....

Avoid the problem by creating the channel outside of the goroutine,
this way both reading and writing work reliably.